% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{preProcessing}
\alias{preProcessing}
\title{Dissolve geometry as a pre-processing step}
\usage{
preProcessing(x, split = TRUE)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{split}{If \code{TRUE} (default), then a single-part-polygon is returned}
}
\value{
Geometry of class \code{sf} with dissolved boundaries
}
\description{
This function performs a \code{sf::st_union} on a given geometry
}
\keyword{erase}
\keyword{feature,}
\keyword{simple}
