% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rgrass_dissolve}
\alias{rgrass_dissolve}
\title{Dissolves geometries using GRASS GIS}
\usage{
rgrass_dissolve(
  x,
  ...,
  split = FALSE,
  check.geom = TRUE,
  stringsAsFactors = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{x}{object of class \code{sf}. Always of type polygon.}

\item{...}{other option for \code{v.dissolve} set into params (e.g. layer or column).}

\item{split}{export vector data as single features. Default: \code{FALSE}}

\item{check.geom}{If set to  \code{TRUE} then geometry is checked with \code{sf::st_is_valid} (\link[sf]{geos_query}). If there are invalid geometries, geometries are repaired using \code{st_make_valid} (\link[sf]{valid}). Default: \code{TRUE}}

\item{stringsAsFactors}{Default: \code{FALSE}}

\item{quiet}{If \code{FALSE} then comments are printed. Default: \code{TRUE}}
}
\value{
Geometry of class \code{sf}
}
\description{
This function dissolves a geometry. Optionnally, based on a field input.
}
\keyword{dissolve}
\keyword{feature,}
\keyword{simple}
