% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rgrass_make_valid}
\alias{rgrass_make_valid}
\title{Make geometries valid using GRASS GIS}
\usage{
rgrass_make_valid(x, tool = "break", check.geom = FALSE,
  stringsAsFactors = FALSE, quiet = TRUE)
}
\arguments{
\item{x}{object of class \code{sf}. Always of type polygon.}

\item{tool}{Cleaning tool. Can be \code{"break, snap, rmdangle, chdangle, rmbridge, chbridge, rmdupl, rmdac, bpol, prune, rmarea, rmline, rmsa"}. Default: \code{"break"}}

\item{check.geom}{If set to \code{TRUE} then geometry is checked with \code{sf::st_is_valid} (\link[sf]{geos_query}). If there are invalid geometries, geometries are repaired using \code{st_make_valid} (\link[lwgeom]{valid}). Default: \code{FALSE}}

\item{stringsAsFactors}{Default: \code{FALSE}}

\item{quiet}{If \code{FALSE} then comments are printed. Default: \code{TRUE}}
}
\value{
Geometry of class \code{sf}
}
\description{
This function cleans a geometry.
}
\keyword{dissolve}
\keyword{feature,}
\keyword{simple}
