% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rgrass_overlay}
\alias{rgrass_overlay}
\title{Overlays two vector geometries using GRASS GIS}
\usage{
rgrass_overlay(
  x,
  y,
  operator = "and",
  ...,
  unique.colnames = TRUE,
  check.geom = TRUE,
  stringsAsFactors = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{x}{object of class \code{sf}. First element: Should be either of type line or polygon}

\item{y}{object of class \code{sf}. Second element: Always of type polygon.}

\item{operator}{operator of \code{v.overlay}. \code{"and"}: intersection,  \code{"or"}: union, \code{"not"}: difference, \code{"xor"}: symmetrical difference. Default: \code{"and"}}

\item{...}{other option for \code{v.overlay} set into params.}

\item{unique.colnames}{Make columns names unique using \link[base]{make.unique}. Default: \code{TRUE}}

\item{check.geom}{If set to  \code{TRUE} then geometry is checked with \code{sf::st_is_valid} (\link[sf]{geos_query}). If there are invalid geometries, geometries are repaired using \code{st_make_valid} (\link[lwgeom]{valid}). Default: \code{TRUE}}

\item{stringsAsFactors}{Default: \code{FALSE}}

\item{quiet}{If \code{FALSE} then comments are printed. Default: \code{TRUE}}
}
\value{
Geometry of class \code{sfc}
}
\description{
This function overlais one geometry with another. The projection must be identical.
}
\keyword{erase}
\keyword{feature,}
\keyword{simple}
