% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rsaga_erase}
\alias{rsaga_erase}
\title{Erase one geometry from another using Saga GIS}
\usage{
rsaga_erase(x, y, method = "1", split = "0", attributes = "1",
  env.rsaga = RSAGA::rsaga.env(), check.geom = TRUE, quiet = TRUE)
}
\arguments{
\item{x}{object of class \code{sf}. First element: Should be either of type line or polygon}

\item{y}{object of class \code{sf}. Second element: Always of type polygon.}

\item{method}{method of erase. Either \code{"1"}: Polygon-Polygon-Erase , or \code{"2"}: Line-Polygon-Erase. Default: \code{"1"}}

\item{split}{Set to \code{"1"}, if multi-part polygons should be splitted to single-part polygons. Default: \code{"0"}}

\item{attributes}{attributes inherited to intersection result. \code{0} polygon, \code{1} line, \code{2} line and polygon. Default: \code{"1"}}

\item{env.rsaga}{SAGA GIS environemnt. Default: \link[RSAGA]{rsaga.env}}

\item{check.geom}{If set to  \code{TRUE} then geometry is checked with \code{sf::st_is_valid} (\link[sf]{geos_query}). If there are invalid geometries, geometries are repaired using \code{st_make_valid} (\link[lwgeom]{valid}). Default: \code{TRUE}}

\item{quiet}{If \code{FALSE} then comments are printed. Default: \code{TRUE}}
}
\value{
Geometry of class \code{sfc}
}
\description{
This function erase one geometry from another. The projection must be identical.
}
\keyword{erase}
\keyword{feature,}
\keyword{simple}
