% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{st_dissolve}
\alias{st_dissolve}
\title{Dissolve geometry}
\usage{
st_dissolve(x, by = NULL, ...)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{by}{field for dissolving. Default: \code{NULL}}

\item{...}{Optional: field statistcs}
}
\value{
dissolved geometry of class \code{sf}
}
\description{
This function dissovle a geometry based on a field. Optionnally, field statistics can be computed.
}
\keyword{dissolve}
\keyword{feature,}
\keyword{simple}
