% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_edge_length.R
\name{st_edge_length}
\alias{st_edge_length}
\title{Calculate edge length between two geometries}
\usage{
st_edge_length(x, y, dist.x = 0, dist.y = 0, geom.boundary = NULL,
  do.preProcessing = TRUE, return.geom = FALSE, quiet = TRUE)
}
\arguments{
\item{x}{polygon of class \code{sf} representing the basis geometry}

\item{y}{polygon of class \code{sf} representing geometry of which the edge length is calculated}

\item{dist.x}{buffer distance in mapping units around geometry using \code{sf::st_buffer}. Default: \code{0}}

\item{dist.y}{buffer distance in mapping units around geometry using \code{sf::st_buffer}. Default: \code{0}}

\item{geom.boundary}{polygon of class \code{sf} representing subregions, e.g. administrative boundaries}

\item{do.preProcessing}{If \code{TRUE} (default), the input of \code{geom.frag} is, first, dissolved to single part feature, and second, splitted to multi-parts}

\item{return.geom}{If set to \code{TRUE}, intermediate geometries are returned as well. Default: \code{FALSE}}

\item{quiet}{If set to \code{FALSE}, actual state is printed to console. Default: \code{TRUE}.}
}
\value{
\code{list} with edge length, and if \code{return.geom} is \code{TRUE} also the edge-length-geometry (and eventually buffer results).
}
\description{
This function calculates the edge length between two geometries.
}
\note{
If the edge line has incomprehensible gabs(can happen for "complicated" geometries), it can help to use a small buffer distance (< 1, e.g. 1e-04) to get a complete edge line.
}
\keyword{edge}
\keyword{length}
