% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_integration_index.R
\name{st_integration_index}
\alias{st_integration_index}
\title{Calculate integration index}
\usage{
st_integration_index(
  tool = "sf",
  geom.old,
  geom.new,
  geom.boundary = NULL,
  tol = 0.1,
  precision = 0,
  env.rsaga = NULL,
  snap.rgrass = 1e-05,
  dist.new = 0,
  override = TRUE,
  return.geom = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{tool}{tool to perform erase-vector-operation. Default is \code{"sf"}. However, for complex polygons \code{"grass"} is highly recommended. Fot the use of \code{"grass"} a valid GRASS GIS-session mus be initiated. \code{"saga"} is also supported, then \code{env.rsaga} must be properly set}

\item{geom.old}{object of class \code{sf} representing a land use class of a previous time step}

\item{geom.new}{object of class \code{sf} representing a land use class of a following time step}

\item{geom.boundary}{polygon of class \code{sf} representing subregions, e.g. administrative boundaries}

\item{tol}{tolerance value for overlapping area meter square}

\item{precision}{precision for process using \code{sf::st_set_precision}. See \link[sf]{st_precision}. Default: \code{0}}

\item{env.rsaga}{environment of \code{SAGA GIS}. If st_erase fails then \code{SAGA GIS} erase is used. Default: \code{NULL}, but in function call if not set: \link[RSAGA]{rsaga.env}}

\item{snap.rgrass}{similar to \code{sf::st_set_precision}. Snapping threshold for boundaries. Default: \code{1e-04}}

\item{dist.new}{buffer distance in mapping units around new geometry (result of erase) using \code{sf::st_buffer}. Default: \code{0}}

\item{override}{If \code{TRUE} projection of geometry-operation is replaced by projection of \code{geom.old}. This flag prevents errors when the original datum is not found by the \code{tool}. Dafault: \code{FALSE}}

\item{return.geom}{If set to \code{TRUE}, intermediate geometries are returned as well. Default: \code{FALSE}}

\item{quiet}{If set to \code{FALSE} actual state is printed to console. Default: \code{TRUE}.}
}
\value{
Vector with integration index (completly integrated: 2/3 < R < 1; good integrated: 1/3 < R < 2/3; low integrated: 0 < R < 1/3; not integrated: 0)
}
\description{
This function calculates an index defining the intergration of new land use area into conisting land use area
of the same land use class conisdering two different time steps.
}
\note{
If the unique border line has incomprehensible gabs(can happen for "complicated" geometries), it can help to use a small buffer distance (< 0.05, e.g. 1e-04) by \code{dist.new} to get a complete edge line
Code is based on the following references:
\itemize{
\item Meinel, G., & Winkler, M. (2002). Spatial analysis of settlement and open land trends in urban areas on basis of RS data œ studies of five European cities over a 50-year period.
\item Siedentop, S., Heiland, S., Lehmann, I., & Schauerte-Lüke, N. (2007). Regionale Schlüsselindikatoren nachhaltiger Flächennutzung für die Fortschrittsberichte der Nationalen Nachhaltigkeitsstrategie–Flächenziele (Nachhaltigkeitsbarometer Fläche). Abschlussbericht, BBR-Forschungen, H, 130.
}
Depending on the selected \code{tool}, the result of the vector-operation can differ significantly!
}
\keyword{edge}
\keyword{feature,}
\keyword{index,}
\keyword{interior}
\keyword{ratio}
\keyword{shape}
\keyword{simple}
