% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{st_make_grid_lines}
\alias{st_make_grid_lines}
\title{Make multiline fishnet}
\usage{
st_make_grid_lines(x, cellsize)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{cellsize}{cell size of fishnet grid in meter}
}
\value{
Geometry of class \code{sfc}
}
\description{
This function creates a fishnet, consisting of multilines in x and y direction.
}
\keyword{feature,}
\keyword{fishnet,}
\keyword{multilines}
\keyword{simple}
