% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_MSI.R
\name{st_MSI}
\alias{st_MSI}
\title{Calculate the mean shape index of a polygons}
\usage{
st_MSI(
  tool = "sf",
  x,
  geom.boundary = NULL,
  field = NULL,
  do.preProcessing = TRUE,
  return.geom = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{tool}{tool to perform dissolve operation. Default is \code{"sf"}. However, for complex polygons \code{"grass"} is highly recommended. Fot the use of \code{"grass"} a valid GRASS GIS-session mus be initiated. \code{"saga"} is NOT supported}

\item{x}{polygon of class \code{sf}}

\item{geom.boundary}{polygon of class \code{sf} representing subregions, e.g. administrative boundaries. Default: \code{NULL}}

\item{field}{field name or table column (\code{character}) on which the shannon index is calculated}

\item{do.preProcessing}{If \code{TRUE} (default), the input of \code{geom.frag} is, first, dissolved by field, and second, splitted to multi-parts. By this step it is assured, that polygon connected to each other are summarized}

\item{return.geom}{If set to \code{TRUE}, geometries (e.g. dissolved) are returned as well. Default: \code{FALSE}}

\item{quiet}{If set to \code{FALSE} actual state is printed to console. Default: \code{TRUE}.}
}
\value{
\code{data.frame} with mean shape index and area weighted mean shape index
}
\description{
This function calculates the mean shape index of polygons.
}
\note{
Code is based on the following references:
\itemize{
\item Forman, R.T.T., & Godron, M. (1986). Landscape Ecology. Cambridge.
\item Siedentop, S., Heiland, S., Lehmann, I., & Schauerte-Lüke, N. (2007). Regionale Schlüsselindikatoren nachhaltiger Flächennutzung für die Fortschrittsberichte der Nationalen Nachhaltigkeitsstrategie–Flächenziele (Nachhaltigkeitsbarometer Fläche). Abschlussbericht, BBR-Forschungen, H, 130.
}
}
\keyword{area}
\keyword{feature,}
\keyword{index}
\keyword{index,}
\keyword{mean}
\keyword{shape}
\keyword{simple}
\keyword{weighted}
