% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_shannon_index.R
\name{st_shannon_index}
\alias{st_shannon_index}
\title{Calculate the Shannon's Diversity Index}
\usage{
st_shannon_index(tool = "sf", x, field, do.preProcessing = TRUE,
  return.geom = FALSE, quiet = TRUE)
}
\arguments{
\item{tool}{tool to perform dissolve operation. Default is \code{"sf"}. However, for complex polygons \code{"grass"} is highly recommended. Fot the use of \code{"grass"} a valid GRASS GIS-session mus be initiated. \code{"saga"} is NOT supported}

\item{x}{polygon of class \code{sf}}

\item{field}{field name or table column (\code{character}) on which the shannon index is calculated}

\item{do.preProcessing}{If \code{TRUE} (default), the input of \code{geom.frag} is, first, dissolved by field, and second, splitted to multi-parts. By this step it is assured, that polygon connected to each other are summarized}

\item{return.geom}{If set to \code{TRUE}, geometries (e.g. dissolved) are returned as well. Default: \code{FALSE}}

\item{quiet}{If set to \code{FALSE} actual state is printed to console. Default: \code{TRUE}.}
}
\value{
\code{list} with overall Shannon's Diversity Index, and index for each landscape classes. If \code{return.geom} is \code{TRUE} than \code{list} with result and geometry on which SHDI is based is returned.
}
\description{
This function calculates the Shannon's Diversity Index for landscapes.
}
\note{
Code is based on the following references:
\itemize{
\item McGarigal, K., & Marks, B. J. (1995). FRAGSTATS: spatial pattern analysis program for quantifying landscape structure. Gen. Tech. Rep. PNW-GTR-351. Portland, OR: US Department of Agriculture, Forest Service, Pacific Northwest Research Station. 122 p, 351.
}
}
\keyword{Diversity}
\keyword{Index}
\keyword{Shannon's}
