% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLSM_data.R
\docType{data}
\name{VLSM_data}
\alias{VLSM_data}
\title{Data-set of VLSM-package.}
\format{
A list with three entries:
\describe{
\item{CLC_12}{CORINE Land Cover of year 2012}
\item{CLC_18}{CORINE Land Cover of year 2018}
\item{ADMIN_DIST}{Administrative boundaries}
}
}
\source{
\itemize{
\item \url{https://www.govdata.de/web/guest/daten/-/details/corine-land-cover-deutschland-25-ha-2018-datensatz}
\item \url{https://www.bkg.bund.de/DE/Produkte-und-Services/Shop-und-Downloads/Digitale-Geodaten/Verwaltungsgebiete-Verwaltungsgrenzen/verwaltungsgebiete_cont.html}
}
}
\usage{
VLSM_data
}
\description{
A data-set containing the CORINE Land Cover (CLC) Germany 25 ha for the two districs Jena and Saale-Holzland.
CLC is from the year 2012 and 2018. Moreover, administrative boundaries are available.
}
\keyword{datasets}
