% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLSM_data_result.R
\docType{data}
\name{VLSM_data_result}
\alias{VLSM_data_result}
\title{Data-set of VLSM-package with results.}
\format{A list with three entries:
\describe{
\item{CLC_12_J}{CORINE Land Cover of year 2012}
\item{CLC_18_J}{CORINE Land Cover of year 2018}
\item{TH12_mesh}{result of effective mesh size for CLC-2012}
\item{TH18_mesh}{result of effective mesh size for CLC-2018}
\item{TH12_urbanSprawl}{result of urban sprawl for CLC-2012}
\item{TH18_urbanSprawl}{result of urban sprawl for CLC-2018}
\item{TH18_integration}{result of integration index}
\item{plot_limit}{plot limits for zoom into integration index geometries}
\item{CLC2012_MSI}{result of mean shape index for CLC-2012}
\item{CLC2018_MSI}{result of mean shape index for CLC-2018}
\item{TH12_EdgeLength}{result of edge length calculation for CLC-2012}
\item{TH18_EdgeLength}{result of edge length calculation for CLC-2018}
\item{TH12_Shannon}{result of Shannon Diversity index for CLC-2012}
\item{TH18_Shannon}{result of Shannon Diversity index for CLC-2018}
}}
\usage{
VLSM_data_result
}
\description{
A data-set containing function results.
}
\keyword{datasets}
