% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLSM_QGIS.R
\docType{data}
\name{VLSM_QGIS}
\alias{VLSM_QGIS}
\title{Data-set of VLSM-package.}
\format{A list with three entries:
\describe{
\item{image_intro}{Image for section 2}
\item{image_mesh}{Image for section 3.1}
\item{image_ffe}{Image for section 3.2}
}}
\usage{
VLSM_QGIS
}
\description{
A data-set containing magick-raw-images for the QGIS-vignette. All images can be read by \code{magick::image_read(...}.
}
\keyword{datasets}
