% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{st_erase}
\alias{st_erase}
\title{Erase one geometry from another}
\usage{
st_erase(x, y, precision = 0, do.subset = TRUE)
}
\arguments{
\item{x}{object of class \code{sf}}

\item{y}{object of class \code{sf}}

\item{precision}{\code{st_set_precision}, see \link[sf]{st_precision}. Default: \code{0}}

\item{do.subset}{Perform \code{st_intersects} (see \link[sf]{geos_binary_pred}) to subset geometry of \code{x}.Default: \code{TRUE}}
}
\value{
Geometry of class \code{sfc}
}
\description{
This function erases one geometry from another. The projection must be identical.
}
\keyword{erase}
\keyword{feature,}
\keyword{simple}
