% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_integration_index.R
\name{st_integration_index}
\alias{st_integration_index}
\title{Calculate integration index}
\usage{
st_integration_index(geom.old, geom.new, geom.boundary = NULL,
  tol = 0.1, precision = 0, env.rsaga = NULL, use.saga = FALSE,
  return.geom = FALSE, quiet = FALSE)
}
\arguments{
\item{geom.old}{object of class \code{sf} representing a land use class of a previous time step}

\item{geom.new}{object of class \code{sf} representing a land use class of a following time step}

\item{geom.boundary}{polygon of class \code{sf} representing subregions, e.g. administrative boundaries}

\item{tol}{tolerance value for overlapping area meter square}

\item{precision}{precision for process using \code{sf::st_set_precision}. See \link[sf]{st_precision}. Default: \code{0}}

\item{env.rsaga}{environment of \code{SAGA GIS}. If st_erase fails then \code{SAGA GIS} erase is used. Default: \code{NULL}, but in function call if not set: \link[RSAGA]{rsaga.env}}

\item{use.saga}{use \code{SAGA GIS} for erase process. Default: \code{FALSE}}

\item{return.geom}{If set to \code{TRUE}, intermediate geometries are returned as well. Default: \code{FALSE}}

\item{quiet}{show output on console. Default: \code{FALSE}}
}
\value{
Vector with integration index (completly integrated: 2/3 < R < 1; good integrated: 1/3 < R < 2/3; low integrated: 0 < R < 1/3; not integrated: 0)
}
\description{
This function calculates an index defining the intergration of new land use area into conisting land use area
of the same land use class conisdering two different time steps.
}
\note{
Code is based on the following references:
\itemize{
\item Meinel, G., & Winkler, M. (2002). Spatial analysis of settlement and open land trends in urban areas on basis of RS data œ studies of five European cities over a 50-year period.
\item Siedentop, S., Heiland, S., Lehmann, I., & Schauerte-Lüke, N. (2007). Regionale Schlüsselindikatoren nachhaltiger Flächennutzung für die Fortschrittsberichte der Nationalen Nachhaltigkeitsstrategie–Flächenziele (Nachhaltigkeitsbarometer Fläche). Abschlussbericht, BBR-Forschungen, H, 130.
}
}
\keyword{edge}
\keyword{feature,}
\keyword{index,}
\keyword{interior}
\keyword{ratio}
\keyword{shape}
\keyword{simple}
