% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_mesh.R
\name{st_mesh}
\alias{st_mesh}
\title{Calculate the effective mesh size}
\usage{
st_mesh(geom.frag, geom.boundary = NULL, total.area = NULL,
  conv = 10000, do.preProcessing = TRUE, return.geom = FALSE,
  quiet = TRUE)
}
\arguments{
\item{geom.frag}{polygon of class \code{sf} representing the fragmentation geometry}

\item{geom.boundary}{polygon of class \code{sf} representing subregions, e.g. administrative boundaries}

\item{total.area}{Numeric value representing size for area. Only to use if \code{geom.boundary} is not present. Value must match with the conversion constant \code{c} (default hectare). \code{Default: NULL}}

\item{conv}{constant to convert original square meter output. Default: \code{10000} to convert to hectare. If set to \code{1}, than meter square is the result.}

\item{do.preProcessing}{If \code{TRUE} (default), the input of \code{geom.frag} is, first, dissolved to single part feature, and second, splitted to multi-parts. By this step it is assured, that polygon connected to each other are summarized}

\item{return.geom}{If set to \code{TRUE}, intermediate geometries are returned as well. Default: \code{FALSE}}

\item{quiet}{If set to \code{FALSE} actual state is printed to console. Default: \code{TRUE}.}
}
\value{
If \code{return.geom} is \code{TRUE} than \code{list} with result and geometry is returned. Otherwhise result is returned in form of a \code{data.frame}
}
\description{
This function calculates the effective mesh size.
}
\keyword{effective}
\keyword{feature,}
\keyword{mesh}
\keyword{mesh,}
\keyword{simple}
\keyword{size}
