% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_shape_indices.R
\name{st_shape_indices}
\alias{st_shape_indices}
\title{Calculate the shape indices of a polygon}
\usage{
st_shape_indices(x)
}
\arguments{
\item{x}{object of class \code{sf}}
}
\value{
\code{List} with shape index and interior edge ratio values
}
\description{
This function calculates the shape index of a polygon.
}
\keyword{edge}
\keyword{feature,}
\keyword{index,}
\keyword{interior}
\keyword{ratio}
\keyword{shape}
\keyword{simple}
