% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_urban_sprawl.R
\name{st_urban_sprawl}
\alias{st_urban_sprawl}
\title{Calculate urban sprawl}
\usage{
st_urban_sprawl(tool = "sf", geom.urban, geom.boundary = NULL,
  dist = c(100, 100), trans = function(x, trans.k) {     x - 1 + 1/(x +
  trans.k) }, trans.k = 1, tol = 0.1, extent = NULL,
  force.extent = FALSE, precision = 0, do.preProcessing = TRUE,
  return.geom = FALSE, env.rsaga = NULL, quiet = TRUE)
}
\arguments{
\item{tool}{tool to perform erase-vector-operation. Default is \code{"sf"}. However, for complex polygons \code{"grass"} is highly recommended. Fot the use of \code{"grass"} a valid GRASS GIS-session mus be initiated. \code{"saga"} is also supported, then \code{env.rsaga} must be properly set.}

\item{geom.urban}{polygon of class \code{sf} representing the fragmentation geometry}

\item{geom.boundary}{polygon of class \code{sf} representing subregions, e.g. administrative boundaries}

\item{dist}{\code{vector} containing distance between lines in \code{x} and \code{y} direction. Default: \code{c(100, 100)} \code{m}}

\item{trans}{transformation function \code{x-1+1/(x+trans.k)} with \code{x} as free line and \code{trans.k} as constant}

\item{trans.k}{constant in \code{km} for transformation function \code{trans}. Default: \code{1}}

\item{tol}{tolerance value for intersection with erased lines. Buffering procedure is used. Default: \code{0.1} \code{m}}

\item{extent}{Numeric value representing extent for area. Format of vector: \code{c(xmin, xmax, ymax, ymin)}. Default: \code{NULL}}

\item{force.extent}{If \code{TRUE} extent is used instead of \code{geom.boundary} (if both are present). Default: \code{FALSE}}

\item{precision}{precision for process \code{sf::st_set_precision}. See \link[sf]{st_precision}. Default: \code{0}}

\item{do.preProcessing}{If \code{TRUE} (default), the input of \code{geom.frag} is, first, dissolved to single part feature, and second, splitted to multi-parts. By this step it is assured, that polygon connected to each other are summarized}

\item{return.geom}{If set to \code{TRUE}, intermediate geometries are returned as well. Default: \code{FALSE}}

\item{env.rsaga}{environment of \code{SAGA GIS}. If \code{st_erase} fails then \code{SAGA GIS erase} is used. Default: \code{NULL}, but in function call if not set: \link[RSAGA]{rsaga.env}}

\item{quiet}{If set to \code{FALSE}, actual state is printed to console. Default: \code{TRUE}.}
}
\value{
strong urban sprawl: 40-50%, less urban sprawl: 80-90%
}
\description{
This function calculates the urban sprawl of urban area.
}
\note{
Code is based on the following references:
\itemize{
\item Ackermann, W., Schweiger, M., Sukopp, U., Fuchs, D., & Sachteleben, J. (2013). Indikatoren zur biologischen vielfalt: Landschaftszersiedlung (Biodiversity indicators. Development and accounting). Naturschutz und Biologische Vielfalt, 132.
}
Depending on the selected \code{tool}, the result of the vector-operation can differ significantly!
}
\keyword{feature,}
\keyword{simple}
\keyword{sprawl}
\keyword{urban}
